+++ clid/editmeta.py
@@ -4,37 +4,31 @@
 
 import os
 
-import stagger
 import npyscreen as npy
 
 from . import base
-from . import _const
+from . import const
+from . import readtag
 
 
 class SingleEditMetaView(base.ClidEditMetaView):
     """Edit the metadata of a *single* track."""
     def create(self):
         file = self.parentApp.current_files[0]
-        try:
-            meta = stagger.read_tag(file)
-        except stagger.NoTagError:
-            meta = stagger.Tag23()   # create a id3v2.3 tag instance
-
+        meta = readtag.ReadTags(file)
         self.filenamebox = self.add(self._title_textbox, name='Filename',
                                     value=os.path.basename(file).replace('.mp3', ''))
         self.nextrely += 2
         super().create()
 
-        for tbox, field in _const.TAG_FIELDS.items():  # show file's tag
-            getattr(self, tbox).value = str(getattr(meta, field))   # str for track_number
+        for tbox, field in const.TAG_FIELDS.items():  # show file's tag
+            getattr(self, tbox).value = getattr(meta, field)   # str for track_number
 
     def get_fields_to_save(self):
-        return _const.TAG_FIELDS
+        return const.TAG_FIELDS
 
-    def on_ok(self):
-        c = super().on_ok()
-        if c is None:
-            return None   # some error like invalid date or track number has occured
+    def do_after_saving_tags(self):
+        """Rename the file if necessary."""
         mp3 = self.files[0]
         new_filename = os.path.dirname(mp3) + '/' + self.filenamebox.value + '.mp3'
         if mp3 != new_filename:   # filename was changed
@@ -42,8 +36,6 @@ class SingleEditMetaView(base.ClidEditMetaView):
             main_form = self.parentApp.getForm("MAIN")
             main_form.value.replace_file(old=mp3, new=new_filename)
             main_form.load_files()
-
-        self.switch_to_main()
 
 
 class MultiEditMetaView(base.ClidEditMetaView):
@@ -55,11 +47,8 @@ class MultiEditMetaView(base.ClidEditMetaView):
 
     def get_fields_to_save(self):
         # save only those fields which are not empty, to files
-        return {tbox: field for tbox, field in _const.TAG_FIELDS.items()\
+        return {tbox: field for tbox, field in const.TAG_FIELDS.items()\
                 if getattr(self, tbox).value}
 
-    def on_ok(self):
-        c = super().on_ok()
-        if c is None:
-            return None   # some error like invalid date or track number has occured
-        self.switch_to_main()
+    def do_after_saving_tags(self):
+        pass   # nothing more to do